requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.5
openpyxl>=3.1.0
anthropic>=0.80.0
python-dotenv>=1.0.0
